Moved phylostylotastic code into its own module.
